major clean up; removing warnings, putting everything in its right place
