+++ src/hevm/src/EVM/Facts.hs
@@ -34,8 +34,9 @@ module EVM.Facts
   ) where
 
 import EVM          (VM, Contract)
-import EVM.Concrete (Word, litWord)
-import EVM          (balance, nonce, storage, bytecode, env, contracts)
+import EVM.Concrete (Word)
+import EVM.Symbolic (litWord)
+import EVM          (balance, nonce, storage, bytecode, env, contracts, contract, state)
 import EVM.Types    (Addr)
 
 import qualified EVM
